Refactor helpers and pages
