+++ src/types.ts
@@ -1,0 +1,5 @@
+export interface CompanyField {
+  field: string;
+  recommended: string;
+  considerations: string;
+}
